% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cmip_data.R
\name{create_crss_cmip_nf_files}
\alias{create_crss_cmip_nf_files}
\title{Create CRSS natural flow input files from CMIP data}
\usage{
create_crss_cmip_nf_files(iFile, oFolder, startYear, endYear = 2060,
  oFiles = CRSSNFInputNames(), overwriteFiles = FALSE)
}
\arguments{
\item{iFile}{Path to netcdf file containing the CMIP based natural inflow
data. See 'Details'.}

\item{oFolder}{Path to the top level directory where the trace folders and
input files will be created.}

\item{startYear}{The year to start the trace files in. Data will be trimmed
to start in this year.}

\item{endYear}{The final year of data the trace files will contain.}

\item{oFiles}{The CRSS natural inflow file names to use for the individual
traces files.}

\item{overwriteFiles}{A boolean that determines whether or not the function
should overwrite existing files. See 'Details'.}
}
\value{
\code{iFile} is invisibly returned as the main purpose of the function is to write
many files.
}
\description{
\code{create_crss_cmip_nf_files()} creates individual trace files for CRSS from
a netcdf file containing CMIP based natural flow data.
}
\details{
\code{create_crss_cmip_nf_files()} will create individual trace files named by the
\code{oFiles} argument for all traces that exist in \code{iFile}. Individual trace
folders, e.g., trace1, trace2, traceN, are created for all traces found in
\code{iFile}. \code{iFile} should be a netcdf file that contains a variable called
\code{naturalFlow}. If it does not, then the function will error. The netcdf file
should contain a 3-dimensional array in (trace, gageNumber, time) format.

The CMIP data currently exists for 1950 - 2099. If the user specifies years
outside of this range, the function will abort. Because CRSS needs data to
start in the year that the CRSS simulations begin, this function trims the
data based on \code{startYear} and \code{endYear}, and correctly formats the trace
files to begin in January of \code{startYear} and end in December of \code{endYear}.

\code{oFiles} sets the individual file names for the natural inflow locations. If
you do not use \code{\link{CRSSNFInputNames}()}, oFiles should contain 29 strings:
one for each of the natural inflow locations, and should be specified in the
default order corresponding to the gages in \code{\link{nfGageNames}()}.

\code{overwriteFiles} allows the user to control whether existing files within the
trace folders should be overwritten (default is they are not).
}
\examples{
\dontrun{
create_crss_cmip_nf_files(
  "cmip5_bcsd.nc",
  oFolder = "c:/model/CRSS/dmi/cmip5",
  startYear = 2018,
  endYear = 2060,
  overwriteFiles = TRUE
)
}

}
