% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createISMMatrix.R
\name{createISMMatrix}
\alias{createISMMatrix}
\title{Create a matrix of data based on the ISM method}
\usage{
createISMMatrix(xtsData, startMonth, nYrs = NA)
}
\arguments{
\item{xtsData}{An xts vector}

\item{startMonth}{The startMonth of the return matrix. Should be able to be
cast to a zoo::yearmon}

\item{nYrs}{The number of years to create the data for. Defaults to the number
of years in xtsData, but can be less.}
}
\value{
xts matrix
}
\description{
Takes in a matrix of historical data and applies the ISM method to it. Because
this is typically used to create future hydrology data, the entire matrix
will have a new start month.
}

