% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createSysCondTable}
\alias{createSysCondTable}
\title{\code{createSysCondTable}}
\usage{
createSysCondTable(zz, yrs)
}
\arguments{
\item{zz}{Full data for all years/traces necessary for creating System Conditions table}

\item{yrs}{Vector of years to process all of the System Conditions}
}
\value{
List with two Data frames: one with the System Conditions for the specified
years including the breakout of Lower Elevation Balancing releases and the other without
the Lower Elevation Balancing breakout
}
\description{
\code{createSysCondTable} creates the standard System Conditions table that is commonly created
from CRSS results. The table reports the percent of traces that that simulate various
system conditions, e.g., Lake Powell operating tiers, through time.
}

