% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssi_change_nf_start_date.R
\name{crssi_change_evap_files}
\alias{crssi_change_evap_files}
\title{Change Start Date of Evap Files}
\usage{
crssi_change_evap_files(nTrace, folder, startDate, NZeros)
}
\arguments{
\item{nTrace}{The number of traces to process.}

\item{folder}{A string with either a relative or absolute path that contains
the trace folders, e.g., C:/CRSS/dmi/DNFSinput}

\item{startDate}{New start date. Should be in 2012-1-31 format.}

\item{NZeros}{is the number of zeros to add to the data that is read in}
}
\description{
Changes the start date in all files in each CRSS_DIR/dmi/evap/trace folder.
}
\details{
For evaporation files used in CRSS that are in CRSS_DIR/dmi/Evap, this
function will change their start date, and add 0s on if necessary to extend
the evaporation rates. The files specific to Powell are skipped.
}
