% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssi_change_nf_file_names.R
\name{crssi_change_nf_file_names}
\alias{crssi_change_nf_file_names}
\title{Rename CRSS Input Files}
\usage{
crssi_change_nf_file_names(iFolder, oFolder, nTrace, fromNames, toNames)
}
\arguments{
\item{iFolder}{Path to the input folder cotaining trace folders.}

\item{oFolder}{Path to the ouptut folder containig trace folders.}

\item{nTrace}{Number of traces to process}

\item{fromNames}{A vector of the file names found in iFolder/traceN.}

\item{toNames}{A vector of the file names to create in oFolder/traceN.}
}
\value{
Nothing is returned from function.
}
\description{
Rename the CRSS natural flow or salt input files.
}
\details{
Because multiple versions of CRSS exist and the inflow locations have had
their names changed in recent years, it is necessary to create files with
different file names. It might be easier to copy existing files instead of
creating files from the source data, as \code{\link[=crssi_create_dnf_files]{crssi_create_dnf_files()}} does.
\code{crssi_change_nf_file_names()} assumes the folders
are structured for CRSS input, e.g., C:/CRSS/dmi/NFSinput/trace1/...
}
\examples{
# load the common old and new natural flow files names included with the 
# CRSSIO package.
\dontrun{
iFolder <- 'C:/CRSS/dmi/NFSinputOrig/'
oFolder <- 'C:/CRSS/dmi/NFSinputNew/'
oldFileNames <- nf_file_names(version = 1)
newFileNames <- nf_file_names(version = 2)
crssi_change_nf_file_names(iFolder, oFolder, 107,oldFileNames, newFileNames)
}

}
