% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssi_change_nf_start_date.R
\name{crssi_change_nf_start_date}
\alias{crssi_change_nf_start_date}
\alias{changeStartDate}
\title{Change the start dates of CRSS input trace files}
\usage{
crssi_change_nf_start_date(nTrace, folder, start_year, startDate)

changeStartDate(nTrace, folder, startDate)
}
\arguments{
\item{nTrace}{The number of traces to process.}

\item{folder}{A string with either a relative or absolute path that contains
the trace folders, e.g., 'C:/CRSS/dmi/DNFSinput'.}

\item{start_year}{The new start_year for the trace files. Should be a single
numeric value.}

\item{startDate}{Deprecated.}
}
\description{
\code{crssi_change_nf_start_date()} changes the start date in all files in each
trace folder. Used to change the start date of all natural flow input files.
This Will change the start date for all files in a folder (flow, salinity,
other).
}
