% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cmip_data.R
\name{crssi_create_cmip_nf_files}
\alias{crssi_create_cmip_nf_files}
\title{Create CRSS natural flow input files from CMIP data}
\usage{
crssi_create_cmip_nf_files(iFile, oFolder, startYear, endYear = 2060,
  scenarioNumber, oFiles = nf_file_names(), overwriteFiles = FALSE)
}
\arguments{
\item{iFile}{Path to netcdf file containing the CMIP based natural inflow
data. See 'Details'.}

\item{oFolder}{Path to the top level directory where the trace folders and
input files will be created. This folder should exist before using this
function.}

\item{startYear}{The year to start the trace files in. Data will be trimmed
to start in this year.}

\item{endYear}{The final year of data the trace files will contain.}

\item{scenarioNumber}{The scenario number used as an identifier in CRSS.
See 'Details.'}

\item{oFiles}{The CRSS natural inflow file names to use for the individual
traces files.}

\item{overwriteFiles}{A boolean that determines whether or not the function
should overwrite existing files. See 'Details'.}
}
\value{
\code{iFile} is invisibly returned as the main purpose of the function is to write
many files.
}
\description{
\code{crssi_create_cmip_nf_files()} creates individual trace files for CRSS from
a netcdf file containing CMIP based natural flow data.
}
\details{
\code{crssi_create_cmip_nf_files()} will create individual trace files named by
the \code{oFiles} argument for all traces that exist in \code{iFile}. Individual trace
folders, e.g., trace1, trace2, traceN, are created for all traces found in
\code{iFile}. \code{iFile} should be a netcdf file that contains a variable called
\code{naturalFlow}. If it does not, then the function will error. The netcdf file
should contain a 3-dimensional array in (trace, gageNumber, time) format.

The CMIP data currently exists for 1950 - 2099. If the user specifies years
outside of this range, the function will abort. Because CRSS needs data to
start in the year that the CRSS simulations begin, this function trims the
data based on \code{startYear} and \code{endYear}, and correctly formats the trace
files to begin in January of \code{startYear} and end in December of \code{endYear}.

\code{scenarioNumber} is specified by the user and is saved to the slot specified
by the \code{"crssio.supplyScenarioSlot"} option. The following maps supply
scenario names to the scenario numbers that should be used here. (See the
CRSS model documentation for more details on the scenario number.) This
function will error if the \code{scenarioNumber} provided is 1, 2, or 3 as these
data are not created with this function.
\itemize{
\item 1 = Observed Resampled
\item 2 = Direct Paleo Resampled
\item 3 = Paleo-conditioned
\item 4 = CMIP3 Downscaled GCM Projected
\item 5 = CMIP5 Downscaled GCM Projected, BCSD downscaling, quantile mapping
secondary bias correction
}

In addition to the scenario number, the trace number, the hydrology
increment, and the Sacramento Year Type index are created for each trace.
The names of these slots are controlled by the \code{"crssio.traceNumberSlot"},
\code{"crssio.hydroIncrement"}, and \code{"crssio.sacYTSlot"} options, respectively.
The Sacramento Year Type index is stored in the netcdf file with the natural
flow files, and is created from a regression of a river gage in the
Sacramento Basin. See the CRSS documentation for more details. Finally,
a README file is created in \code{oFolder} that provides some metadata on the
creation of the trace files.

\code{oFiles} sets the individual file names for the natural inflow locations. If
you do not use \code{\link[=nf_file_names]{nf_file_names()}}, oFiles should contain 29 strings:
one for each of the natural inflow locations, and should be specified in the
default order corresponding to the gages in \code{\link[=nf_gage_names]{nf_gage_names()}}.

\code{overwriteFiles} allows the user to control whether existing files within the
trace folders should be overwritten (default is they are not).
}
\examples{
\dontrun{
crssi_create_cmip_nf_files(
  "cmip5_bcsd.nc",
  oFolder = "c:/model/CRSS/dmi/cmip5",
  startYear = 2018,
  endYear = 2060,
  overwriteFiles = TRUE
)
}

}
