% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssi_create_hist_nf_xlsx.R
\name{crssi_create_hist_nf_xlsx}
\alias{crssi_create_hist_nf_xlsx}
\title{Create the HistoricalNaturalFlows.xlsx file}
\usage{
crssi_create_hist_nf_xlsx(modelStartYear, nYearAvg = 5, oFolder = ".")
}
\arguments{
\item{modelStartYear}{The year the CRSS model run should begin.}

\item{nYearAvg}{The number of years to average when filling the intervening
natural flow data.}

\item{oFolder}{The location to save the xlsx file.}
}
\value{
The filename that is created is invisibly returned. The function
creates the HistoricalNaturalFlows.xlsx file, which contains three
worksheets: README, Intervening Natural Flow, Total Natural Flow.
}
\description{
\code{crssi_create_hist_nf_xlsx()} creates the HistoricalNaturalFlows.xlsx that
CRSS relies on. This file needed to be manually updated before each new
model start date before; this function takes care of the manual update.
}
\details{
HistoricalNaturalFlows.xlsx will contain monthly total natural flow above
Lees Ferry for the entire historical natural flow record that exists. It will
also contain monthly intevening natural flow for all gages at and below
Hoover. For these gages, data are appended from the end of the natural flow
record through the year before the CRSS model run starts (\code{modelStartYear}).
The appended data use an
n-year average (\code{nYearAvg}) of the available historical natural flow. For
example, if the CRSS run begins in January 2019 and historical natural flow
exists from 1906 - 2015, then January 2016 - December 2018 will be filled
with the monthly average for 2011-2015 (if \code{nYearAvg == 5}).

The \code{\link[CoRiverNF]{CoRiverNF}} package is relied on for the
historical natural flows.

The filename defaults to HistoricalNaturalFlows.xlsx, but can be modified by
changing the \code{crssio.histNfFile} option.
}
\examples{
# create the file for a CRSS model run that begins in January 2019 and uses
# a 10-year average
\dontrun{
crssi_create_hist_nf_xlsx(2019, nYearAvg = 10)
}
   
}
