% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFInputNames.R
\name{CRSSNatSaltInputNames}
\alias{CRSSNatSaltInputNames}
\title{Get vector of CRSS natural salt input file names}
\usage{
CRSSNatSaltInputNames(version = 2)
}
\arguments{
\item{version}{The CRSS version number. Current version of CRSS is 2. Valid
versions are 1 or 2.}
}
\value{
Vector of characters (file names).
}
\description{
\code{CRSSNatSaltInputNames} returns a vector with the CRSS natural salt file
names in the correct node order.
}
\details{
This function returns a vector with the file names used by CRSS to import
the natural salt. The order matches the node order used by Reclamation and
CRSS and thus should not be modified. As the structure to CRSS changes,
the file names also may change. If this occurs in the future, the function
will be modified to return different values for different versions of CRSS.
}
\examples{
fileNames <- CRSSNatSaltInputNames(2)
}
\seealso{
\code{\link{CRSSNFInputNames}}
}
