% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CRSSNFInputNames}
\alias{CRSSNFInputNames}
\title{Get vector of CRSS natural inflow file names.}
\usage{
CRSSNFInputNames()
}
\value{
Vector of characters (file names).
}
\description{
\code{CRSSNFInputNames} returns a vector with the CRSS natural inflow file
names in the correct node order.
}
\details{
This function returns a vector with the file names used by CRSS to import
the natural flows. The order matches the node order used by Reclamation and
CRSS and thus should not be modified. As the strucutre to CRSS changes,
the file names also may change. If this occurs in the future, the function
will be modified to return different values for different versions of CRSS.

This is the default function used by \code{\link{createCRSSDNFInputFiles}}
for creating the input files for CRSS.
}
\examples{
fileNames <- CRSSNFInputNames()
createCRSSDNFInputFiles('NaturalFlows1906-2012_withExtensions_1.8.15.xlsx','NFSinput/','2015-1-31',50,fileNames)
}
\seealso{
\code{\link{createCRSSDNFInputFiles}}
}

