% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sys_cond_table.R
\name{crsso_get_sys_cond_table}
\alias{crsso_get_sys_cond_table}
\alias{createSysCondTable}
\title{Create standard CRSS system conditions table}
\usage{
crsso_get_sys_cond_table(zz, yrs)

createSysCondTable(zz, yrs)
}
\arguments{
\item{zz}{Full data for all years/traces necessary for creating System
Conditions table. \code{zz} should be a data frame returned from
\code{RWDataPlot::\link[RWDataPlyr]{getDataForAllScens}} that contains all
of the 17 variables necessary to create the system conditions table.}

\item{yrs}{Vector of years to provide the system conditions for.
Ex: `2017:2020``}
}
\value{
Named list with two matrices. The first matrix (\code{'fullTable'})
includes the system conditions for the specified years including the
breakout of Lower Elevation Balancing releases.  The second matrix
(\code{"limitedTable"}) includes the system conditions without the Lower
Elevation Balancing breakout.
}
\description{
\code{crsso_get_sys_cond_table()} creates the standard system conditions table
that is commonly created from CRSS results, e.g., slide 6 at
\url{https://www.usbr.gov/lc/region/g4000/crss-5year.pdf}. The table reports
the percent of traces that simulate various system conditions, e.g., Lake
Powell operating tiers, through time.
}
\examples{
# use RWDataPlyr package to get the data to create the system conditions table
require(RWDataPlyr)
slotAggList <- RWDataPlyr::createSlotAggList(CRSSIO::sysCondSALMatrix())
scenFolder <- 'DNF,CT,IG'
scenName <- 'DNF Hydrology'
scenPath <- system.file('extdata','Scenario/',package = 'RWDataPlyr')
sysData <- RWDataPlyr::getDataForAllScens(scenFolder, scenName, slotAggList,
                                          scenPath, 'tmp.feather', TRUE)
sysCondTable <- crsso_get_sys_cond_table(sysData, 2018:2022)
sysCondTable[['limitedTable']]

}
\seealso{
\code{\link{sysCondSALMatrix}}
}
