% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_to_storage.R,
%   R/storage_to_elevation.R
\name{elevation_to_storage}
\alias{elevation_to_storage}
\alias{elevation2Volume}
\alias{storage_to_elevation}
\title{Convert Between Elevation and Volume for a Reservoir}
\usage{
elevation_to_storage(elevation, reservoir)

elevation2Volume(elevation, reservoir)

storage_to_elevation(storage, reservoir)
}
\arguments{
\item{elevation}{The elevation (in feet) that will be converted to volume.}

\item{reservoir}{The reservoir to convert between elevation and storage.
Spaces are removed from the name, and case is ignored. This should be a
single reservoir.}

\item{storage}{The volume (in acre-feet) that will be converted to an
elevation (in feet).}
}
\value{
\code{elevation_to_storage()} returns volume in acre-feet with the same
length as \code{elevation}, or \code{NA} if the elevation is invalid for the
reservoir.

\code{storage_to_elevation}() returns an elevation in feet with the same
length as \code{storage}, or \code{NA} if the storage is invalid for the reservoir.
}
\description{
\code{elevation_to_storage()} computes the storage (volume) for a given elevation
for a particular reservoir included in CRSS.

\code{storage_to_elevation()} computes the elevation from a given storage (volume)
for a particular reservoir included in CRSS.
}
\details{
Elevation is assumed to be in feet, and volume is assumed to be in acre-feet
for both input and return values in these functions. Currently, the functions
will work for Flaming Gorge, Navajo, Blue Mesa, Powell, and Mead reservoirs.
The elevation-volume tables contain data for 0.5 foot increments. The data
are linearly interpolated for sub 0.5 foot values.

If the input values (\code{elevation} or \code{storage}) are beyond the range of the
resevoir, the function returns \code{NA}.
}
\examples{
elevation_to_storage(1075, "mead")
elevation_to_storage(6000, "Flaming Gorge")
elevation_to_storage(c(6000, 6001, 6004.6), "flamingGorge")

storage_to_elevation(6500000, "mead")
storage_to_elevation(100000, "Flaming Gorge")
storage_to_elevation(c(100000, 101000, 125000), "flamingGorge")

}
