% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get5YrTable.R
\name{get5YearTable}
\alias{get5YearTable}
\title{\code{get5YearTable}}
\usage{
get5YearTable(scenPath, scen, oPath, nYrs)
}
\arguments{
\item{scenPath}{Path to scenario folder}

\item{scen}{The scenario name}

\item{oPath}{Path to write a text file to}

\item{nYrs}{The number of years to create the table for}
}
\value{
Nothing. But two txt files are written out.
}
\description{
\code{get5YearTable} is depreciated and will be removed in future versions. Please combine
\code{\link[RWDataPlot]{getDataForAllScens}} with \code{\link{createSysCondTable}} instead of
using this function.
}

