% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssInput_helperFunctions.R
\name{getAllISMMatrices}
\alias{getAllISMMatrices}
\title{returns all ISM matrices for all nodes}
\usage{
getAllISMMatrices(nfMat, startMonth, nYrs)
}
\arguments{
\item{nfMat}{Matrix of natural flow data}

\item{startMonth}{The startMonth of the return matrix. Should be able to be
cast to a zoo::yearmon}

\item{nYrs}{The number of years to create the data for.}
}
\value{
list of matrices. Each node is one matrix entry into the list
}
\description{
Given a matrix of natural flow data, applies the ISM method to all nodes
}
\keyword{internal}

