% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ism.R, R/ism_get_matrix.R
\name{ism}
\alias{ism}
\alias{ism.crssi}
\alias{ism.crss_nf}
\alias{ism.nfd}
\alias{ism.xts}
\alias{ism.matrix}
\alias{ism_get_site_matrix}
\title{Apply the Index Sequential Method}
\usage{
ism(x, n_years_keep = NA, ...)

\method{ism}{crssi}(x, n_years_keep = NA, ...)

\method{ism}{crss_nf}(x, n_years_keep = NA, ...)

\method{ism}{nfd}(x, n_years_keep = NA, ...)

\method{ism}{xts}(x, n_years_keep = NA, ...)

\method{ism}{matrix}(x, n_years_keep = NA, is_monthly, ...)

ism_get_site_matrix(xtsData, startMonth, nYrs = NA, monthly = TRUE)
}
\arguments{
\item{x}{An R object. \code{ism()} is implemented for matrices, \link[xts:xts]{xts}, \link{nfd},
\link{crss_nf}, and \link{crssi} objects.}

\item{n_years_keep}{The number of years of data to keep. This must be <= the
number of years of data that exists in \code{x}.}

\item{...}{Further arguments passed to subsequent methods.}

\item{is_monthly}{Boolean. \code{TRUE} if data are monthly. \code{FALSE} if data are
annual. See \emph{Details}.}

\item{xtsData}{An xts vector.}

\item{startMonth}{The start month and year of the return matrix. Should be
able to be cast to a \link[zoo:yearmon]{zoo::yearmon}.}

\item{nYrs}{The number of years to create the data for. Defaults to the
number of years in xtsData, but can be less.}

\item{monthly}{Boolean that should be set to \code{TRUE} if the data are monthly;
should set to \code{FALSE} if annual data.}
}
\value{
\code{ism()} returns an object of the same class as \code{x}, but with more
traces, and potentially trimmed to the number of years specified by
\code{n_years_keep}.

\code{ism_get_site_matrix()} returns an \code{xts} matrix with the number of
years/months specified by \code{nYrs} and the number of columns equal to the
number of years in \code{xtsData}
}
\description{
\code{ism()} applies the Index Sequential Method to a single trace of data to
create multiple traces. \code{ism()} will work on matrices, \link[xts:xts]{xts}, \link{nfd},
\link{crss_nf}, and \link{crssi} objects. \code{ism_get_site_matrix()} is a deprecated
precursor to this function that only works on \code{xts} objects.
}
\details{
\code{ism()} will determine if the data are annual or monthly (or contain both for
objects inheriting from \code{nfd}), and will act accordingly. The time step can
be forced by using \code{is_monthly}, but that is only necessary for matrices.
If \code{is_monthly} is used with an \code{nfd} type object that contains annual and
monthly data, an error will occur.

\code{ism_get_site_matrix()} is a deprecated version of \code{ism()} that only works
on \code{xts} objects. It takes in 1 column of historical data for a single
site and applies (ISM) to it. This function allows you to change the start
date of the returned data in it, while \code{ism()} does not. When using \verb{ism(),  [reindex()] should be used after it to change the start date. }ism_get_site_matrix()\verb{can be used on monthly or annual data. If applying it to monthly data, then}xtsData\verb{needs to be monthly data, and}monthly\verb{should be set to}TRUE\verb{. If using annual data, then }xtsData\verb{should  be annual, i.e., all with a December time stamp, and}monthly\verb{should be set to}FALSE\verb{. If }monthly\code{is}FALSE\code{and}xtsData` is
monthly data, an error will occur.
}
\examples{
# monthly data, that will create a 48x4 xts matrix
t1 <- xts::xts(1:48, zoo::as.yearmon("Jan 2000") + seq(0,47)/12)
ism(t1)

# annual data that will create a 5 x 6 matrix
t2 <- xts::xts(1:6, zoo::as.yearmon("Dec 2000") + 0:5)
ism(t2, n_years_keep = 5)

}
