% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFInputNames.R
\name{nf_file_names}
\alias{nf_file_names}
\alias{natsalt_file_names}
\alias{nf_gage_names}
\alias{nf_gage_abbrv}
\alias{CRSSNFInputNames}
\alias{CRSSNatSaltInputNames}
\alias{nfGageNames}
\alias{nfShortNames}
\title{Get vector of natural flow and salt gage names, filenames, and abbreviations}
\usage{
nf_file_names(version = 2)

natsalt_file_names(version = 2)

nf_gage_names()

nf_gage_abbrv()

CRSSNFInputNames(version = 2)

CRSSNatSaltInputNames(version = 2)

nfGageNames()

nfShortNames()
}
\arguments{
\item{version}{The CRSS version number. Current version of CRSS is 2. Valid
versions are 1 or 2.}
}
\value{
Vector of characters with 29 entries (file names, gage names, gage
abbreviations).
}
\description{
These functions return the natural flow and salt file names (\code{nf_file_names()}
and \code{natsalt_file_names()}), the natural flow gage names (\code{nf_gage_names()}), and
the abbreviations (\code{nf_gage_abbrv()}) of the natural flow gage names in the
standard node order used by Reclamation and expected by CRSS.
}
\details{
\code{nf_file_names()} and \code{natsalt_file_names()} return file names that
CRSS is expecting to read in for natural flow and salt input data.

\code{nf_gage_names()} returns a vector with the USGS gage names used by
CRSS  corresponding to the natural flow basins. The order matches the node
order used by Reclamation and CRSS and thus should not be modified.

\code{nf_gage_abbrv()} returns an abbreviated shorthand name for the natural flow
gages/nodes, because it is it is desirable to have shorthand names for many
purposes, i.e., variable names.
}
\examples{
fileNames <- nf_file_names()
\dontrun{
iFiles <- 'NaturalFlows1906-2012_withExtensions_1.8.15.xlsx'
crssi_create_dnf_files(iFile,'NFSinput/','2015-1-31',50,fileNames)
}

# get the gage name for node 20 
nf_gage_name()[20]
# and its shorthand name
nf_gage_abbrv()[20]
# and then see the CRSS natural inflow file name corresponding to this gage
nf_file_names()[20]

}
\seealso{
\code{\link{crssi_create_dnf_files}}
}
