% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_to_annual.R
\name{nf_to_annual}
\alias{nf_to_annual}
\alias{nf_to_annual.xts}
\alias{nf_to_annual.nfd}
\alias{nf_to_annual.crss_nf}
\alias{nf_to_annual.crssi}
\title{Sum monthly natural flow data to annual data}
\usage{
nf_to_annual(x, ...)

\method{nf_to_annual}{xts}(x, ..., year = "cy", full_year = TRUE)

\method{nf_to_annual}{nfd}(x, ..., full_year = TRUE, recompute = FALSE, keep_monthly = TRUE)

\method{nf_to_annual}{crss_nf}(x, ..., full_year = TRUE, recompute = FALSE, keep_monthly = TRUE)

\method{nf_to_annual}{crssi}(x, ..., recompute = FALSE)
}
\arguments{
\item{x}{An object inheriting from \code{xts}.}

\item{...}{Other parameters passed to methods.}

\item{year}{"cy" or "wy" to sum over the calendar or water year,
respectively. For \code{nfd} like objects, year is determined from the year
specified in the object.}

\item{full_year}{Only return sums for full years when \code{TRUE}. Otherwise, will
sum all months in a year, even if that's a partial year.}

\item{recompute}{If \code{nfd} object already has annual data, should the annual
data be recomputed. An error will post if it has annual data and
\code{recompute} is \code{FALSE}.}

\item{keep_monthly}{If \code{TRUE} the monthly data are kept in the returned
object, otherwise they are dropped.}
}
\description{
\code{nf_to_annual()} sums \code{nfd}, and \code{xts} data from monthly data to annual data.
}
\examples{
# can sum monthly data to annual and get the existing stored annual data
library(CoRiverNF)
ann <- nf_to_annual(monthlyTot)
all.equal(ann, cyAnnTot, check.attributes = FALSE)

# for nfd objects, annual data will be added to object and the monthly 
# data are kept by default
nf <- nfd(
  monthlyTot["2000/2002"], 
  flow_space = "total", 
  time_step = "monthly"
)

nf2 <- nf_to_annual(nf)

# nf2 now has annual data, and monthly data

nf2 <- nf_to_annual(nf, keep_monthly = FALSE)
# nf2 no longer has monthly data

}
\seealso{
\code{\link[=nf_to_intervening]{nf_to_intervening()}}, \code{\link[=nf_to_total]{nf_to_total()}}
}
