% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_to_total.R
\name{nf_to_total}
\alias{nf_to_total}
\alias{nf_to_total.crss_nf}
\alias{nf_to_total.nfd}
\alias{nf_to_total.xts}
\title{Convert intervening flow to total flow}
\usage{
nf_to_total(x, ...)

\method{nf_to_total}{crss_nf}(x, keep_intervening = TRUE, recompute = FALSE, ...)

\method{nf_to_total}{nfd}(x, keep_intervening = TRUE, recompute = FALSE, ...)

\method{nf_to_total}{xts}(x, ...)
}
\arguments{
\item{x}{An object inheriting from \link[xts:xts]{xts::xts}, \link{nfd}, \link{crss_nf}, or \link{crssi}.}

\item{keep_intervening}{When \code{FALSE}, the intervening flow are removed from
the object. Otherwise those data remain in the returned object.}

\item{recompute}{If \code{x} already has total flow, the function will fail with
an error when \code{FALSE}. If \code{TRUE}, then the total flow will be recomputed
and the existing data will be overwritten.}
}
\value{
An object of the same class as \code{x}, except when \code{x} is either a
\code{crss_nf} obect or \code{crssi} object and \code{keep_intervening} is \code{FALSE}. See
\strong{Details}.
}
\description{
\code{nf_to_total()} converts intervening flow to total flow for the Colorado
River natural flow network. The functions only work if the objects have the
29 Colorado River natural flow sites, and that dimension's names match
\code{\link[=nf_gage_abbrv]{nf_gage_abbrv()}}. \code{xts}, \code{nfd}, \code{crss_nf}, and \code{crssi} objects
can all have their intervening natural flow converted to total flow. For
\code{xts} objects, the flow is converted without any additional checks, i.e.,
it is assumed the user passed in intervening flow. For the other three types
of objects, they must have intervening data already. If the objects have
both annual and monthly data, both will be converted to total flow.
}
\details{
When \code{keep_intervening} is \code{FALSE} and \code{x} is either a \code{crss_nf} object or
a \code{crssi} object, then the returned object will be an \code{nfd} object, and a
warning will post. This is because \code{crss_nf} and \code{crssi} objects must have
monthly intervening natural flow stored in them.
}
\examples{
# start with an nfd object with intervening annual data
nf_int <- nfd(
  CoRiverNF::cyAnnInt, 
  n_sites = 29, 
  flow_space = "intervening", 
  time_step = "annual"
)

# convert to total
nf <- nf_to_total(nf_int)

# It matches the total natural flow that exists in the CoRiverNF package 
# (except it doesn't have the sheet name attribute)
all.equal(
  zoo::coredata(CoRiverNF::cyAnnTot),
  zoo::coredata(
    nfd_get_trace(nf, 1, flow_space = "total", time_step = "annual")
  )
)

}
