% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfa.R
\name{nfa}
\alias{nfa}
\title{Natural Flow Array (nfa)}
\usage{
nfa(
  data = NA,
  dim = NA,
  flow_space = c("intervening", "total"),
  start_yearmon = NA
)
}
\arguments{
\item{data}{If \code{NA} or of length 1, creates an object with dimensions based
on \code{dim}. Otherwise, creates an object based on the provided data. Data
should be a matrix, array, list, data.frame, or \verb{[xts]} object.}

\item{dim}{\code{NA} or a length 2 numeric vector. If \code{NA} and \code{data} is \code{NA},
then creates an object with one month and one trace of data with the values.
If \code{data} is only one value, then creates based on the user specified
number of months and number of traces. \code{c(months, traces)}. Should be \code{NA}
if data is not NA or length of 1.}

\item{flow_space}{Data are intervening or total flow.}

\item{start_yearmon}{Start year and month of data. If \code{NA}, then assumes the
data begins in January of the current year. Should be a \link[zoo:yearmon]{zoo::yearmon}
object or coercable to one.}
}
\description{
\code{nfa()} creates a natural flow array from the specified \code{data}. Stores data
for multiple months and traces (sequences) for all 29 natural flow sites
(\code{\link[=nf_gage_names]{nf_gage_names()}}).
}
\details{
\emph{Array:} Arrays should be an m x t x 29 array, where m is the number of
months, and t is the number of
traces. If there are rownames, then they must be in "yyyy-mm" format,
otherwise an error will post. Rownames are not required, and if they are not
provided will be set starting with the specified \code{start_yearmon} or assuming
a starting date of January of the current year. The colnames are not required
but will be used in their provided form if they are specified. If they are
not specified, they will be named Trace1 - TraceN. The names of the 3rd
dimension should match \code{\link[=nf_gage_abbrv]{nf_gage_abbrv()}} if provided. If they are not
provided, then assume that the depths match the order in \code{\link[=nf_gage_abbrv]{nf_gage_abbrv()}}.
}
