% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_extract.R
\name{nfd_extract}
\alias{nfd_extract}
\title{Extract subsets from nfd objects}
\usage{
nfd_extract(x, i, j, k, l, m)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}.}

\item{i}{The times to extract. Can be numeric or ISO-8601 style character
range.}

\item{j}{Traces to extract. Numeric.}

\item{k}{Sites to extract. Numeric or names of sites.}

\item{l}{The flow space. "intervening", "total", or
\code{c("total", "intervening")}.}

\item{m}{The time step to get. "annual", "monthly", or
\code{c("annual", "monthly")}.}
}
\value{
\code{nfd_extract()} returns an \code{nfd} object if \code{x} is an \code{nfd} object.
If \code{x} is a \code{crss_nf} object, a \code{crss_nf} object is returned, unless the
extraction results in an invalid \code{crss_nf} object, i.e., there are less than
the required 29 sites and/or the time_step or flow_space no longer includes
monthly intervening data.
}
\description{
Extract subsets from \code{nfd} (and \code{crss_nf}) objects by time, trace, site,
flow space, and/or timestep dimensions.
}
\details{
When calling \code{nfd_extract()} not all dimensions have to be specified. When
dimensions are not specified, all data for that dimension are returned.

Times (\code{i}) can be specified by row index or ISO-8601 style character ranges.
For example: \code{"2000/2002"} would return all data for 2000-2002. See
\code{\link[xts]{[.xts}} for more details.
}
