% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_overlapping_ts.R, R/reindex.R
\name{has_overlapping_ts}
\alias{has_overlapping_ts}
\alias{reindex}
\title{Helper Functions for Natural Flow Data Time Steps}
\usage{
has_overlapping_ts(x, exact = TRUE)

reindex(x, start_year)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}, \link{crss_nf}, or \link{crssi}.}

\item{exact}{Boolean. Do the annual and monthly data have to overalp
"exactly". See \emph{Details.}}

\item{start_year}{The new starting year as a string or numeric.}
}
\description{
The objects that store natural flow data (\link{nfd}, \link{crss_nf}, or \link{crssi}) all
include multiple xts objects. These functions help modify or query the
timestep component of these objects.
}
\details{
\code{has_overlapping_ts()} determines whether the the annual and monthly data
within the object are overlapping. If \code{exact} is \code{TRUE}, the start and end
timestep for the monthly and annual data must match "exactly". This means for
an object storing calendar year data that the monthly data will start in
January, year1 and end in December, year2 while the annual data must start
in December, year1 and end in December, year2. The analagous is true for
water year data, except the start and end months are October and September,
respectively.

If the object contains only monthly, or only annual data,
\code{has_overlapping_ts()} will always return TRUE.

\code{reindex()} changes the timesteps, i.e., indeces, of natural flow data as
stored in \link{nfd}, \link{crss_nf}, or \link{crssi} objects.
}
