% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfdplot.R
\name{print.nfdplot}
\alias{print.nfdplot}
\alias{save_nfdplot}
\title{Print and save nfdplots}
\usage{
\method{print}{nfdplot}(x, ...)

save_nfdplot(nfdplot, filename, width = NA, height = NA)
}
\arguments{
\item{x, nfdplot}{An object inheriting from \code{nfdplot}.}

\item{...}{Other parameters not used by this method.}

\item{filename}{File name to create on disk. Should be a .png or .pdf file.
If it is a .png file, then it will create multiple files with \verb{\%02d}
inserted at the end of the file to ensure unique file names. Directory
should already exist.}

\item{width, height}{Plot size in inches. If not specified, defaults to 7.}
}
\value{
\code{x} and \code{nfdplot} are ivisbily returned.

Invisibly returns \code{nfdplot}
}
\description{
\code{save_nfdplot()} is a convenient function for saving \code{nfdplots}. While
\code{print.nfdplot()} prints covers showing the plots when R is in interactive
mode. \code{nfdplot} objects are returned by \code{\link[=plot.nfd]{plot.nfd()}}, \code{\link[=plot.nfdstats]{plot.nfdstats()}},
and \code{\link[=plot.nfdcdf]{plot.nfdcdf()}}.
}
