% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFInputNames.R
\name{nfGageNames}
\alias{nfGageNames}
\title{Get vector of CRSS natural inflow USGS gage names}
\usage{
nfGageNames()
}
\value{
Vector of characters (file names).
}
\description{
\code{nfGageNames} returns a vector with the CRSS natural flow node USGS gage
names in the correct node order.
}
\details{
This function returns a vector with the USGS gage names used by CRSS corresponding to the
the natural flow basins. The order matches the node order used by Reclamation and
CRSS and thus should not be modified.

The gage names returned here, match the CRSS natural inflow slot names returned by 
\code{\link{CRSSNFInputNames}}
}
\examples{
# get the gage name for node 20 
nfGageNames()[20]
# and then see the CRSS natural inflow slot name corresponding to this gage
CRSSNFInputNames()[20]
}
\seealso{
\code{\link{CRSSNFInputNames}}
}

