% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFInputNames.R
\name{nfShortNames}
\alias{nfShortNames}
\title{Get vector of CRSS natural inflow shorthand names}
\usage{
nfShortNames()
}
\value{
Vector of characters (file names).
}
\description{
\code{nfShortNames} returns a vector with the CRSS natural flow node 
shorthand names in the correct node order.
}
\details{
For many purposes, it is desirable to have shorthand names for the natural flow nodes
used in CRSS. This function returns a vector with the shorthand names used by CRSS corresponding to the
the natural flow basins. The order matches the node order used by Reclamation and
CRSS and thus should not be modified.

The shorthand names returned here, match the CRSS natural inflow slot names returned by 
\code{\link{CRSSNFInputNames}}
}
\examples{
# get the gage name for node 20 
nfGageNames()[20]
# and its shorthand name
nfShortNames()[20]
# and then see the CRSS natural inflow slot name corresponding to this gage
CRSSNFInputNames()[20]
}
\seealso{
\code{\link{CRSSNFInputNames}}, \code{\link{nfGageNames}}
}
