% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-meko2017.R, R/data-woodhouse.R
\docType{data}
\name{meko2017}
\alias{meko2017}
\alias{woodhouse}
\title{Paleo reconstructions of Lees Ferry flow}
\format{
\code{meko2017} contains an \link{nfd} object containing Meko et al. 2017's
"most skillful" reconstruction of Lees Ferry data for water years
1416-2015.

\code{woodhouse} contains an \link{nfd} object containing Woodhouse et al.
2006's "Lees-B" reconstruction. This includes Lees Ferry flow reconstructed
for water years 1490-1998.
}
\source{
\href{https://pdfs.semanticscholar.org/fbaf/487604e2537c8f51037035e59c13113edeeb.pdf}{Meko, D. M., C. A. Woodhouse and E. R. Bigio, (2017), "Southern California Tree-Ring Study," California Department of Water Resources}

\href{https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2005WR004455}{Woodhouse, C.A., S.T. Gray, and D.M. Meko, 2006. Updated streamflow reconstructions for the Upper Colorado River basin.  Water Resources Research.}
}
\usage{
meko2017

woodhouse
}
\description{
There are various paleo reconstructions of Lees Ferry flow. Two of these
data sets are included in this package for ease of use in CRSS: Woodhouse
et al. (2006) and Meko et al. (2017).
}
\keyword{datasets}
