% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_disag.R
\name{paleo_disagg}
\alias{paleo_disagg}
\title{Spatial and temporal diaggregattion of paleo flow data}
\usage{
paleo_disagg(x, ann_flw, mon_flw, nsite, nsim, ofolder = NULL)
}
\arguments{
\item{x}{The annual paleo data to disaagregate.}

\item{ann_flw}{Observed annual flow data used for picking analog year.}

\item{mon_flw}{Intervening monthly natural flow. Used for spatially and
temporaly disaggregating paleo data based on the index year.}

\item{nsite}{The number of sites to diaggregate the data too.}

\item{nsim}{Number of times to repeat the space/time disaggregation.}
}
\description{
This is a direct resampling of LB tribs, i.e., no scale factor is applied to
them.
}
\examples{
\dontrun{
# read in annual synthetic mon_flw for disag
x <- matrix(scan("data-raw/Meko.txt"), ncol = 2, byrow = TRUE) 
# intervening natural flow mon_flw - monthly CY text file
mon_flw <- matrix(
  scan("data-raw/CRB_CY_MON_IV_29_06_08.txt"), 
  ncol = 29, 
  byrow = TRUE
)

# observed annual flow for picking analog disag yr
ann_flw <- matrix(scan("data-raw/LF_06_08.txt"), ncol = 2, byrow = TRUE)
zz <- paleo_disagg(x, ann_flw, mon_flw, 29, 1)
}

}
\references{
Nowak, Kenneth, James Prairie, Balaji Rajagopalan, Upmanu Lall.
(2010) \emph{A nonparametric stochastic approach for multisite disaggregation of
annual to daily streamflow.} Water Resources Research.
}
\author{
Ken Nowak
}
