% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nfd.R
\name{plot.nfd}
\alias{plot.nfd}
\title{Plot nfd objects}
\usage{
\method{plot}{nfd}(
  x,
  trace = -1,
  site = 1,
  flow_space = "both",
  time_step = "both",
  base_units = NULL,
  which = "box",
  show = TRUE,
  ...
)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}.}

\item{trace}{The trace(s) to use as a numeric vector. If \code{-1}, then all
traces are used, otherwise the traces specified in the vector are used.}

\item{site}{The site to plot as a scalar numeric or character. Should be a
site number, or a site name.}

\item{flow_space}{The flow space to plot. "both", "intervening", or "total".}

\item{time_step}{The time step of data to plot. "both", "annual", or
"monthly".}

\item{base_units}{The units of the data. Used for the y-axis label.}

\item{which}{The type of plot to create. Can be multiple plot types, but
must be "box", "cloud", "spaghetti", or some combination of these three
types. Can also specify numerically as 1, 2, and 3, respectively. See the
\emph{Annual plots} and \emph{Monthly plots} sections for details.}

\item{show}{Boolean. Should the plots be shown if in interactive mode. If
\code{TRUE}, the user can enter through all returned plots.}

\item{...}{Other parameters not used by this method.}
}
\value{
\code{nfdplot} object. See \code{\link[=print.nfdplot]{print.nfdplot()}}, \code{\link[=save_nfdplot]{save_nfdplot()}}.
}
\description{
Plot \link{nfd} objects (including \link{crss_nf} and \link{crssi} objects) for a single
site. Plotting these
objects are meant to give a cursory look at the underlying data. For more
complex statistics see \code{\link[=nfd_stats]{nfd_stats()}}.
}
\section{Annual plots}{

Annual plots include boxplots, cloud plots, and spaghetti plots; each plot
shows the annual values across multiple traces through time. Box and cloud
plots use the 5th, 25th, 75th, and 95th percentiles as box and whiskers (in
the box plot) and as the different shaded regions (in the cloud plot).
The spaghetti plots show every single trace as a line plot.
}

\section{Monthly plots}{

Monthly plots only include boxplots by month. These boxplots aggregate data
for all months and all traces. The months are ordered from January - December
if the underlying object has a calendar year (cy) year attribute, and
October - September if it has a water year (wy) year attribute.
}

