% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssInput_helperFunctions.R
\name{readAndFormatNFExcel}
\alias{readAndFormatNFExcel}
\title{Read in NF Excel file and format it}
\usage{
readAndFormatNFExcel(iFile)
}
\arguments{
\item{iFile}{Relative or absolute file path to the Excel file}
}
\value{
An xts matrix beginning in January 1906 with 29 columns.
}
\description{
Read in the NF Excel file, removes whitespace and unnecessary rows and columns.
}
\keyword{internal}
