% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_vrain_nf_calc.R
\name{st_vrain_nf_calc}
\alias{st_vrain_nf_calc}
\title{Compute St. Vrain Natural Flow for CRSS}
\usage{
st_vrain_nf_calc(cy_vol = NULL, co_tot_nf = NULL)
}
\arguments{
\item{cy_vol}{xts object that has annual
(calendar year), total natural flow for GlenwoodSprings. Columns are assumed
to be traces and rows are years.}

\item{co_tot_nf}{\link{nfd} object that has annual, total data and has the
"cy" year type. It also must have at least Glenwood Springs volumes.
Volumes at other locations are ignored.}
}
\value{
xts object containing calendar year total natural flow for the St.
Vrain. Columns are traces and rows are years.
}
\description{
\code{st_vrain_nf_calc()} estimates annual (calendar year) natural flow in
the St. Vrain river in Lyons, CO as a function of total natural flow
at Glenwood Springs.
}
\details{
CRSS uses St. Vrain natural flow to estimate the volume of transmountain
diversions (see CRSS object 'TMD East Slope Supply').

St. Vrain natural flow can be generated from the Colorado River natural flow
data since these must always be available for use in CRSS. St. Vrain natural
natural flow is estimated using a linear regression created by Reclamation
and total natural flow at Glenwood Springs. See
https://github.com/BoulderCodeHub/CRSSIO/tree/master/data-raw/StVrain_NaturalFlow_Documentation
for details on the linear regression.
}
\examples{
st_vrain_nf_calc(
  xts::xts(
    matrix(sample(CoRiverNF::cyAnnTot$GlenwoodSprings,15), nrow = 3, ncol = 5), 
    order.by = zoo::as.yearmon("Dec 2000") + 0:2
  )
)

nf <- nfd(CoRiverNF::cyAnnTot, flow_space = "total", 
  time_step = "annual",
  year = "cy")

st_vrain_nf_calc(co_tot_nf = nf)

}
