% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimCCNFFiles.R
\name{trimCCNFFiles}
\alias{trimCCNFFiles}
\title{Trim Climate Change Natural Flow Files}
\usage{
trimCCNFFiles(startYear, endYear, iFolder, nTraces = 112, force = FALSE)
}
\arguments{
\item{startYear}{The desired start year. Should be after 1950 and before 2099.}

\item{endYear}{The desired end year. Should be after 1950 and before 2099.}

\item{iFolder}{The path to the trace files, e.g., '/dmi/VIC/'.}

\item{nTraces}{The number of traces to process. The default is 112, which is
the number of traces in the CMIP3 climate change hydrology.}

\item{force}{Boolean. If \code{TRUE} the function will trim the files as specified,
but if it is \code{FALSE}, then the function stops without trimming any files.}
}
\value{
The number of files that were processed
}
\description{
\code{trimCCNFFiles()} trims the climate change natural flow input data to
start and/or end in a different year.
}
\details{
The climate change natural flows were developed to start in January 1950.
Unlike the other natural flow files used in CRSS, the data are associated
with specific years, i.e., the index sequential method is not used on the
climate change hydrology. Sometimes it is necessary to trim the data to
start, and end in a particular year. The function will return data starting
in January of the \code{startYear} and ending in December of \code{endYear}.

While the function is typically used on climate change natural flow files, it
will still work with other natural flow input files.
}
\note{
Assumes folder numbers will always start at 1. Additionally, it only attempts
to trim files that have expected natural flow or natural salt file names.
These are the file names returned by \code{\link{CRSSNFInputNames}()} and
\code{\link{CRSSNatSaltInputNames}()}.
}
\examples{
# Trim all 112 traces found in 'CRSS/dmi/VIC' to start in Jan-2017 and 
# end in Dec-2019
\dontrun{
trimCCNFFiles(2017,2019,'CRSS/dmi/VIC/')
}

}
