% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nfd.R
\name{write_nfd}
\alias{write_nfd}
\title{Create Natural Flow Files for Each Site}
\usage{
write_nfd(x, path, overwrite = FALSE)
}
\arguments{
\item{x}{\link{nfd} or \link{crss_nf} object.}

\item{path}{Directory to save files to.}

\item{overwrite}{Boolean. If \code{TRUE} and files already exists in
\code{path}/traceN, the files will be overwritten.}
}
\description{
\code{write_nfd()} creates csv files for each site, flow type, and time step.
In the directory provided, there will be one folder for each flow type-time
step combination, and it each of those folders there will a csv files for
each site. The csv files will be named based on the site names in the \link{nfd}
object. Each csv is in time by trace format
}
\details{
Folders will be created inside path e.g.:
| path/
| monthly_intervening/
| site1.csv
| site2.csv
| ...
| annua_total/
| site1.csv
| site2.csv
| ...
}
