% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssInput_helperFunctions.R
\name{writeNFFilesByNode}
\alias{writeNFFilesByNode}
\title{Write out all of the trace files for a given node}
\usage{
writeNFFilesByNode(nfXts, oFile, oFolder, headerInfo)
}
\arguments{
\item{nfXts}{Natural flow data as a matrix for a single node. The number of
columns is the number of traces that will be written out to
oFolder/trace\link{n}/oFile}

\item{oFile}{File name that each file will be saved as}

\item{oFolder}{Folder location to write files to}
}
\description{
Write out all of the trace files for a given node
}
\keyword{internal}
