% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssInput_helperFunctions.R
\name{writeSingleFile}
\alias{writeSingleFile}
\title{write out a single trace of data}
\usage{
writeSingleFile(xData, fPath, headerInfo)
}
\description{
write out a single trace of data
}
\keyword{internal}
