% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crssInput_helperFunctions.R
\name{writeTraceSupplyNumbers}
\alias{writeTraceSupplyNumbers}
\title{write out the trace number and supply scenario number, to a given trace folder
folderPath should be the top level folder, e.g., dmi/NFSinput}
\usage{
writeTraceSupplyNumbers(traceNum, supplyScenNum, folderPath)
}
\description{
write out the trace number and supply scenario number, to a given trace folder
folderPath should be the top level folder, e.g., dmi/NFSinput
}
\keyword{internal}

