% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copyAndChangeNFFileNames}
\alias{copyAndChangeNFFileNames}
\title{Rename CRSS Input Files}
\usage{
copyAndChangeNFFileNames(iFolder, oFolder, nTrace, fromNames, toNames)
}
\arguments{
\item{iFolder}{Path to the input folder cotaining trace folders.}

\item{oFolder}{Path to the ouptut folder containig trace folders.}

\item{nTrace}{Number of traces to process}

\item{fromNames}{A vector of the file names found in iFolder/traceN.}

\item{toNames}{A vector of the file names to create in oFolder/traceN.}
}
\value{
Nothing is returned from function.
}
\description{
Rename the CRSS natural flow and salt input files.
}
\details{
Because multiple versions of CRSS exist and the inflow locations have had their names changed
in recent years, it is necessary to create files with different file names. It might be
easier to copy existing files instead of creating files from the source data, as
\code{\link{createCRSSDNFInputFiles}} does. \code{copyAndChangeNames} assumes the folders
are structured for CRSS input, e.g., C:/CRSS/dmi/NFSinput/trace1/NFInput.Files.
}
\examples{
# load the common old and new natural flow files names included with the CRSSIO package.
data(newNFNames)
data(oldNFNames)
copyAndChangeNames('C:/CRSS/dmi/NFSinputOrig/','C:/CRSS/dmi/NFSinputNew/',107,oldNFNames$Slot, newNFNames$Slot)
}

