% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createCRSSDNFInputFiles}
\alias{createCRSSDNFInputFiles}
\title{Create CRSS Natural Flow Input Files}
\usage{
createCRSSDNFInputFiles(iFile, oFolder, startDate, simYrs,
  oFiles = CRSSNFInputNames())
}
\arguments{
\item{iFile}{The relative or absolute path to the excel workbook.}

\item{oFolder}{The location to write out all trace files.}

\item{startDate}{The start date to be listed in each trace file; should be in 2014-1-31 format}

\item{simYrs}{The number of years data each file should contain. Should be
no longer than the number of years in the input data.}

\item{oFiles}{A matrix of the file names (input into CRSS). The default uses
\code{\link{CRSSNFInputNames}}. This must be specified in the correct order, i.e.,
the same order as the nodes in the input Excel file.}
}
\value{
Nothing is returned by the function, but it writes out many files.
}
\description{
Creates the CRSS natural flow input files from the standard excel workbook that is used to publish
the natural flow data.
}
\details{
\code{createCRSSDNFInputFiles} creates individual trace files for the observed
natural flow on the Colorado River Basin using the Index Sequential Method
(INSERT REFERENCE). Trace files are formated and saved in a format expected
by the Colorado River Simulation System (CRSS). Data is read from the natural
flow workbook available at \url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html}.
}
\examples{
# will create 50 years for 107 traces based on the 1906-2012 record
creatCRSSDNFInputFiles('NaturalFlows1906-2012_withExtensions_1.8.15.xlsx','NFSinput/','2015-1-31',50)
}
\seealso{
\code{\link{CRSSNFInputNames}}
}

