% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createISMMatrix.R
\name{createISMMatrix}
\alias{createISMMatrix}
\title{Create a matrix of data based on the ISM method}
\usage{
createISMMatrix(xtsData, startMonth, nYrs = NA, monthly = TRUE)
}
\arguments{
\item{xtsData}{An xts vector}

\item{startMonth}{The startMonth of the return matrix. Should be able to be
cast to a zoo::yearmon}

\item{nYrs}{The number of years to create the data for. Defaults to the number
of years in xtsData, but can be less.}

\item{monthly}{\code{TRUE} if the data is monthly data; \code{FALSE} if it is annual data}
}
\value{
xts matrix with the number of years/months specified by \code{nYrs} 
and the number of columns equal to the number of years of data in \code{xtsData}
}
\description{
Takes in a matrix of historical data and applies the ISM method to it. Because
this is typically used to create future hydrology data, the entire matrix
will have a new start month.
}
\details{
The method can be used on monthly or annual data. If you are applying it to 
monthly data, then \code{xtsData} needs to be monthly data, and \code{monthly}
should be set to \code{TRUE}. If using annual data, then \code{xtsData} should 
be annual, i.e., all with a December timestamp, and \code{monthly} should be
set to \code{FALSE}. If \code{monthly} is \code{FALSE} and \code{xtsData} is 
monthly data, an error will occur.
}
\examples{
# monthly data, that will create a 48x4 xts matrix
t1 <- xts::xts(1:48, zoo::as.yearmon("Jan 2000") + seq(0,47)/12)
createISMMatrix(t1, "Jan 2020")

# annual data that will create a 5 x 6 matrix
t2 <- xts::xts(1:6, zoo::as.yearmon("Dec 2000") + 0:5)
createISMMatrix(t2, "Dec 2020", nYrs = 5, monthly = FALSE)

}
