% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCRSSInput_addIn.R
\name{crss_input_addin}
\alias{crss_input_addin}
\title{RStudio Addin to Create CRSS Input Files}
\usage{
crss_input_addin()
}
\description{
\code{crss_input_addin()} is accessed via a RStudio Addin. It is a GUI
(Shiny Gadget) used to set key parameters of \code{\link[=crssi_create_dnf_files]{crssi_create_dnf_files()}}
before running it.
}
\details{
To use the Addin, RStudio v0.99.878 or later must be used. The key user input
to \code{createCRSSDNFInputFiles} can be set in the GUI. The \code{oFiles}
arguement uses the default value of \code{\link{CRSSNFInputNames}}.
}
\seealso{
\code{\link[=crssi_create_dnf_files]{crssi_create_dnf_files()}}, \code{\link[=nf_file_names]{nf_file_names()}}
}
