\name{CRSSIO-package}
\alias{CRSSIO-package}
\alias{CRSSIO}
\docType{package}
\title{
Package to manage the input/output of CRSS data.
}
\description{
Provides functions to generate input files commmonly used by the Colorado River Simulation System (CRSS) from publicly available data, as well as functions to modify these files. Figures that are commonly used to QA/QC and undersand CRSS results are also automated with this package.
}
\details{
\tabular{ll}{
Package: \tab CRSSIO\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-10\cr
License: \tab CCO\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
R. Alan Butler

Maintainer: R. Alan Butler \url{https://github.com/rabutler/CRSSIO/issues}
}
\references{
 ISM method
}
\keyword{ package }
\seealso{
\code{\link[xlsx:read.xlsx]{read.xlsx}}
}
\examples{
~~ simple examples of the most important functions ~~
}
