% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRSSIO.R
\docType{package}
\name{CRSSIO-package}
\alias{CRSSIO}
\alias{CRSSIO-package}
\title{CRSSIO: Package to Manage the Input and Output of CRSS Data}
\description{
Provides functions to generate input files commmonly used by the
Colorado River Simulation System (CRSS) from publicly available data, as
well as functions to modify these files. Figures that are commonly used to
QA/QC and undersand CRSS results are also automated with this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BoulderCodeHub/CRSSIO}
  \item Report bugs at \url{https://github.com/BoulderCodeHub/CRSSIO/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{r.alan.butler+CRSSIO@gmail.com}

}
\keyword{internal}
