% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRSSIO.R
\docType{package}
\name{CRSSIO-package}
\alias{CRSSIO}
\alias{CRSSIO-package}
\title{CRSSIO: Package to Manage the Input and Output of CRSS Data}
\description{
Provides functions to generate input files commmonly used by the
Colorado River Simulation System (CRSS) from publicly available data, as
well as functions to modify these files. Figures that are commonly used to
QA/QC and undersand CRSS results are also automated with this package.
}
\section{Package Options}{

The first four options dictate file names that are created by
\code{\link[=crssi_create_dnf_files]{crssi_create_dnf_files()}} and \code{\link[=crssi_create_cmip_nf_files]{crssi_create_cmip_nf_files()}}:
\describe{
\item{\code{crssio.supplyScenarioSlot}}{The file name (slot name) used for the
supply scenario number. Default: \code{"HydrologyParameters.SupplyScenario"}}
\item{\code{crssio.traceNumberSlot}}{The file name (slot name) used for the
trace number. Default: \code{"HydrologyParameters.TraceNumber"}}
\item{\code{crssio.hydroIncrement}}{The file name (slot name) used for the
hydrology increment value. Default:
\code{"MeadFloodControlData.hydrologyIncrement"}}
\item{\code{crssio.sacYTSlot}}{The file name (slot name) used for the
Sacramento year type index. Default: \code{"MWD_ICS.SacWYType"}}
}

Additional options for file I/O:
\describe{
\item{\code{crssio.histNfFile}}{The file name that will be created by
\code{crssi_create_hist_nf_xlsx()}. Default: \code{"HistoricalNaturalFlow.xlsx"}}
\item{\code{crssio.nf_sheet_name}}{The sheet name to read from the natural flow
Excel file if using Excel instead of the CoRiverNF package in
\code{\link[=crssi_create_dnf_files]{crssi_create_dnf_files()}}. Default: \code{"InterveningNaturalFlow"}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BoulderCodeHub/CRSSIO}
  \item Report bugs at \url{https://github.com/BoulderCodeHub/CRSSIO/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov}

}
\keyword{internal}
