% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation2Volume.R
\name{elevation_to_storage}
\alias{elevation_to_storage}
\alias{elevation2Volume}
\title{Compute Volume from Elevation for a Reservoir}
\usage{
elevation_to_storage(elevation, reservoir)

elevation2Volume(elevation, reservoir)
}
\arguments{
\item{elevation}{The elevation (in feet) that will be converted to volume}

\item{reservoir}{The reservoir to compute the conversion for. Spaces are
removed from the name, and case is ignored.}
}
\value{
Volume in acre-feet, or \code{NA} if the elevation is invalid for the
reservoir.
}
\description{
\code{elevation_to_storage()} computes the storage (volume) for a given elevation
for a particular reservoir included in CRSS.
}
\details{
Elevation is assumed to be provided in feet, and the volume will be returned
in acre-feet. Currently, the function will work for Flaming Gorge, Navajo,
Blue Mesa, Powell, and Mead reservoirs. The elevation-volume tables contain
data for 0.5 foot increments. Volume is linearly interpolated for sub 0.5
foot values of elevation.

If the elevation is beyond the range of the resevoir, the function returns
\code{NA}.
}
\examples{
elevation_to_storage(1075, "mead")
elevation_to_storage(6000, "Flaming Gorge")
elevation_to_storage(c(6000, 6001, 6004.6), "flamingGorge")

}
