% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ism.R
\name{ism}
\alias{ism}
\alias{ism.crssi}
\alias{ism.crss_nf}
\alias{ism.nfd}
\alias{ism.xts}
\alias{ism.matrix}
\title{Apply the Index Sequential Method}
\usage{
ism(x, n_years_keep = NA, ...)

\method{ism}{crssi}(x, ...)

\method{ism}{crss_nf}(x, ...)

\method{ism}{nfd}(x, ...)

\method{ism}{xts}(x, ...)

\method{ism}{matrix}(x, is_monthly, n_years_keep = NA)
}
\arguments{
\item{x}{An R object. \code{ism()} is implemented for matrices, \link{xts}, \link{nfd},
\link{crss_nf}, and \link{crssi} objects.}

\item{n_years_keep}{The number of years of data to keep. This must be <= the
number of years of data that exists in \code{x}.}

\item{...}{Further arguments passed to subsequent methods.}

\item{is_monthly}{Boolean. \code{TRUE} if data are monthly. \code{FALSE} if data are
annual.}
}
\description{
\code{ism()} applies the index sequential method to one site (column) of data.
}
