% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd.R
\name{nfd}
\alias{nfd}
\title{If data is a scalar, then n_time_steps and n_trace are used to determine the
number of traces and time steps. Otherwise, those values are assumed from
the dimensions of the data and the data class.}
\usage{
nfd(
  data = NA,
  start_yearmon = NA,
  n_months = NA,
  n_trace = 1,
  flow_space = c("intervening", "total", "both"),
  time_step = c("annual", "monthly", "both"),
  year = c("cy", "wy")
)
}
\description{
For initializing blank annual data - the number of years is computed as the
number of full years with a minimum of 1 year of data:
\code{n_years = max(floor(n_months / 12), 1)}. The timestep will begin in
December of the year as specified in \code{start_yearmon} if \code{year} is calendar
year, or September if the the \code{year} is water year.
}
