% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd.R, R/crss_nf.R
\name{nfd}
\alias{nfd}
\alias{as_nfd}
\alias{is_nfd}
\alias{crss_nf}
\alias{as_crss_nf}
\alias{is_crss_nf}
\title{Natural Flow Data (nfd)}
\usage{
nfd(
  data = NA,
  n_months = NA,
  n_trace = 1,
  n_sites = NA,
  flow_space = c("total", "intervening", "both"),
  time_step = c("annual", "monthly", "both"),
  start_yearmon = NA,
  year = c("cy", "wy"),
  site_names = NA
)

as_nfd(x, ...)

is_nfd(x)

crss_nf(
  data = NA,
  n_months = NA,
  n_trace = 1,
  flow_space = c("intervening", "both"),
  time_step = c("monthly", "both"),
  start_yearmon = NA,
  year = c("cy", "wy")
)

as_crss_nf(x, ...)

is_crss_nf(x)
}
\arguments{
\item{data}{If \code{NA} or of length 1, creates an object with dimensions based
on \code{n_months}, \code{n_trace}, \code{flow_space}, and \code{time_step}. Otherwise, creates
an object based on the provided data. Data should be a matrix, array, or
\verb{[xts]} object.}

\item{n_months}{The number of months. Scalar numeric.}

\item{n_trace}{The number of traces. Scalar numeric.}

\item{n_sites}{The number of sites. Scalar numeric. When NA, the number of
sites will be solved for from the \code{data} shape.}

\item{flow_space}{Data are intervening or total flow (or both). If both, then
will store/create total and intervening flow data.}

\item{time_step}{Data are annual or monthly (or both). If both, then will
store/create annual and monthly data.}

\item{start_yearmon}{Start year and month of data. If \code{NA}, then assumes the
data begins in January of the current year. Should be a \link[zoo:yearmon]{zoo::yearmon}
object or coercable to one. Scalar character.}

\item{year}{Data are calendar year (\code{"cy"}) or water year (\code{"wy"}). This has
implications for the timestep the annual data are stored in (December for
cy and September for wy) as well as how summation functions will be applied
to monthly data.}

\item{site_names}{The names of the sites. If specified, must be the same
length as the number of sites (\code{n_sites}).}

\item{x}{An \code{R} object.}

\item{...}{Other parameters passed to \code{crss_nf()}.}
}
\value{
\code{nfd()} and \code{as_nfd()} return an object of class \code{nfd}.

\code{is_nfd()} returns \code{TRUE} if class inherits from \code{nfd}.

\code{crss_nf()} and \code{as_crss_nf()} return an object of class \code{crss_nf}.

\code{is_crss_nf()} returns \code{TRUE} if class inherits from \code{crss_nf}.
}
\description{
\code{nfd()} creates an object to store natural flow data from the specified
\code{data}. Conceptually, the \code{nfd} object is a 5-dimensional array, where the
first three dimensions represent time (months or years), traces, and sites.
The next two dimensions are the "flow space" (intervening or total flow), and
then the time time step. This allows for the storage of all natural flow data
used in CRSS and Colorado River modeling in one object, or to limit it and
only store the site or flow space of interest.

\code{crss_nf()} is the constructor to create \code{crss_nf} obects. This class is an
extension of \code{nfd}, however it imposes a strict number and naming convention
to the sites. Where \code{nfd} objects can have an arbitrary number of sites, and
they do not have to be named, \code{crss_nf} objects must have 29 sites, the names
must match the expected site names for CRSS gages \code{\link[=nf_gage_abbrv]{nf_gage_abbrv()}}, and the
data must include monthly intervening data. (Annual and/or monthly total can
also be included, but they do not have to be.)
}
\details{
If \code{data} is a scalar, then \code{n_months} and \code{n_trace} are used to determine
the number of traces and time steps. Otherwise, those values are assumed from
the dimensions of the data and the data class. See \strong{Data Types} section for
a description of how different data types are handled.

For initializing blank annual data - the number of years is computed as the
number of full years with a minimum of 1 year of data:
\code{n_years = max(floor(n_months / 12), 1)}. The timestep will begin in
December of the year as specified in \code{start_yearmon} if \code{year} is calendar
year, or September if the the \code{year} is water year.

The data are assumed to always exist for all sites first, and then the number
of timesteps or traces are determined after that.

When creating \code{crss_nf} objects, arrays, matrices, and xts objects must have
the site dimension named and those names must match the expected natural flow
site names \code{\link[=nf_gage_abbrv]{nf_gage_abbrv()}}.
}
\section{Data Types}{


This section describes how different data types for the specified \code{data}/\code{x}
are treated when created the \code{nfd} object.

\emph{Array:} Arrays should be an m x t x s array, where m is the number of
months (or years), t is the number of traces, and s is the number of sites.
Array can also be an m x t x s x 2 array, where \code{x[,,,1]} is total flow
and \code{x[,,,2]} is intervening flow.
If there are rownames, then they must be in "yyyy-mm" format,
otherwise an error will post. Rownames are not required, and if they are not
provided will be set starting with the specified \code{start_yearmon} or assuming
a starting date of January of the current year. Colnames are not required or
used.
The names of the 3rd dimension represent the site names, and will be used if
provided, otherwise the site names remain \code{NULL}.

\emph{Matrix:} Matrices are assumed to be structured as m x s, where m is the
number of months or years and s is the number of sites for a single trace.
The user needs to specify the \code{time_step} and \code{flow_space}, although they are
assumed to be annual total if not specified. If \code{start_yearmon} is not
specified, it is assumed to be January of the current year. The colnames of
the matrix are used for the site names, if provided, otherwise they remain
\code{NULL}.

\emph{xts:} xts objects work the same way as matrices, except that \code{start_yearmon}
is determined form the xts object.

All other data types will result in an error.
}

