% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_extract.R
\name{nfd_extract}
\alias{nfd_extract}
\title{Extract subsets of nfd objects}
\usage{
nfd_extract(x, i, j, k, l, m)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}.}

\item{i}{The times to extract. Can be numeric or ISO-8601 style character
range.}

\item{j}{Traces to extract. Numeric.}

\item{k}{Sites to extract. Numeric or names of sites.}

\item{l}{The flow space. "intervening", "total", or
\code{c("total", "intervening")}.}

\item{m}{The time step to get. "annual", "monthly", or
\code{c("annual", "monthly")}.}
}
\value{
\code{nfd_extract()} always returns an \code{nfd} object.
}
\description{
Extract subsets of \code{nfd} objects by time, trace, site, flow space, and/or
timestep dimensions.
}
\details{
When calling \code{nfd_extract()} not all dimensions have to be specified. When
dimensions are not specified, all data for that dimension are returned.

Times (\code{i}) can be specified by row index or ISO-8601 style character ranges.
For example: \code{"2000/2002"} would return all data for 2000-2002. See
\code{\link[xts]{[.xts}} for more details.
}
