% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_stats.R
\name{nfd_stats}
\alias{nfd_stats}
\title{Compute Basic Statistics on NFD Objects}
\usage{
nfd_stats(x, site, flow_space, time_step, plot = FALSE)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}.}

\item{site}{The site to compute the stats for. Numeric or name (character).}

\item{flow_space}{"total" or "intervening".}

\item{time_step}{"monthly" or "annual".}

\item{plot}{Boolean. If \code{TRUE} will show the plot. Otherwise, user can
subsequently call \code{plot()} on the returned data.}
}
\value{
\code{nfd_stats} data.frame.
}
\description{
\code{nfd_stats()} computes six basic statistics for each trace in any \link{nfd} type
objects. The basic statistics are mean, variance, minimum, maximum, lag-1
correlation, and skew. The resulting statistics are stored in a special
data.frame - a \code{nfd_stats} object that inherits from \code{tbl_df} and
\code{data.frame}. The resulting statistics can be plotted by calling \code{plot()} on
the resulting object.
}
\seealso{
\link{nfd}, \code{\link[=plot.nfd_stats]{plot.nfd_stats()}}
}
