% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_stats.R, R/nfd_pdf.R
\name{nfd_stats}
\alias{nfd_stats}
\alias{nfd_pdf}
\alias{nfd_pdf_get_breaks}
\title{Compute Statistics on NFD Objects}
\usage{
nfd_stats(x, site, flow_space, time_step, plot = FALSE)

nfd_pdf(x, site, flow_space, time_step, which, breaks = NULL, plot = FALSE)

nfd_pdf_get_breaks(nfdpdf)
}
\arguments{
\item{x}{An object inheriting from \link{nfd}.}

\item{site}{The site to compute the stats for. Numeric or name (character).}

\item{flow_space}{"total" or "intervening".}

\item{time_step}{"monthly" or "annual".}

\item{plot}{Boolean. If \code{TRUE} will show the plot. Otherwise, user can
subsequently call \code{plot()} on the returned data.}

\item{which}{The month to compute the cdf for. A vector including any/all
values in \code{1:12}. Ignored if \code{time_step} is annual. Will be sorted if
passed in out of order.}

\item{breaks}{Specifies how the breaks are computed in \code{\link[stats:density]{stats::density()}},
which is used by \code{nfd_pdf()}.
If \code{breaks} is \code{NULL}, then the number of breaks defaults to 25. If
\code{breaks} is a single number, that is the number of breaks that will be
used. Otherwise, \code{breaks} specifies the breaks, i.e., the number of breaks,
and the minimum (\code{from}) and maximum (\code{to}) breaks that are used in
\code{stats::density()}. If specifiying for monthly data, \code{breaks} should be a
matrix where there is a column for each month that a cdf will be computed
for (\code{which}).}

\item{nfdpdf}{An object inheriting from \code{nfd_pdf}.}
}
\value{
\code{nfd_stats} data.frame.

\code{nfd_pdf} data.frame.

\code{nfd_pdf_get_breaks()} returns a numeric vector (for annual data) or
matrix (for monthly data).
}
\description{
\code{nfd_stats()} computes six basic statistics for each trace in any \link{nfd} type
objects. \code{nfd_pdf()} computes an empirical PDF for each trace and returns the
probabilities and magnitudes associated with the PDF.
}
\details{
The basic statistics are mean, variance, minimum, maximum, lag-1
correlation, and skew. The resulting statistics are stored in a special
data.frame - a \code{nfd_stats} object that inherits from
\code{data.frame}. The resulting statistics can be plotted by calling \code{plot()} on
the resulting object.

If using the output from \code{nfd_pdf()} to then specify the breaks for another
call to \code{nfd_pdf()}, it is highly recommended to use
\code{nfd_pdf_get_breaks()}. \code{nfd_pdf_get_breaks()} will return a matrix of all
the breaks if \code{nfdpdf} contains monthly data, which is the format expected
in \code{nfd_pdf()}. For annual data, it still works but is less necesary.
}
\seealso{
\link{nfd}, \code{\link[=plot.nfd_stats]{plot.nfd_stats()}}, \code{\link[=plot.nfd_pdf]{plot.nfd_pdf()}}
}
