% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_xx_nfd.R
\name{as.data.frame.nfd}
\alias{as.data.frame.nfd}
\alias{as.data.frame.crssi}
\title{Convert \code{nfd} objects to other classes}
\usage{
\method{as.data.frame}{nfd}(x, wide = TRUE, ...)

\method{as.data.frame}{crssi}(x, wide = TRUE, ...)
}
\arguments{
\item{x}{Object inheriting from \code{nfd}.}

\item{wide}{When \code{TRUE}, the sites are included as columns. When \code{FALSE} the
sites are converted to a \code{site} column, making the data.frame longer.}

\item{...}{Other parameters passed to \code{\link[base:as.data.frame]{base::as.data.frame()}}}
}
\value{
\code{data.frame} with columns for \code{date}, \code{trace}, \code{time_step}, and
\code{flow_type}. When \code{wide == TRUE}, there is another column for each site.
When \code{wide == FALSE}, \code{site}, and \code{value} columns. If \code{x} is a \code{crssi}
object, than the returned data.frame will also include Sacramento Year Type
Index data as another column (\code{wide == TRUE}) or as another entry into
\code{site} (\code{wide == FALSE}).
}
\description{
\code{as.data.frame} converts \code{nfd}, \code{crss_nf}, and \code{crssi} objects to
\code{data.frame}s.
}
