% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfd_stats.R
\name{plot.nfd_stats}
\alias{plot.nfd_stats}
\title{Plot nfd_stats Objects}
\usage{
\method{plot}{nfd_stats}(x, points = NULL, base_units = NULL, show = TRUE, ...)
}
\arguments{
\item{x}{An object inheriting from \code{nfd_stats}.}

\item{points}{An object inheriting from \code{nfd_stats}. Optional. Stats in this
object are shown as points instead of boxplots.}

\item{base_units}{Optional. Used as y-axis label.}

\item{show}{Boolean. If \code{TRUE} and in interactive mode, will show the plot
in the plot window.}

\item{...}{Additional options passed to \code{\link[=geom_point]{geom_point()}}. \code{size}, \code{shape}, and
\code{color} can be overridden.}
}
\value{
\code{nfdplot} object.
}
\description{
\code{plot.nfd_stats()} plots the resulting "base" statistics after calling
\code{\link[=nfd_stats]{nfd_stats()}}. Statistics are plotted as boxplots across traces. Historical
or some other reference statistics can be shown using \code{points}.
}
\seealso{
\code{\link[=nfd_stats]{nfd_stats()}}, \code{\link[=save_nfdplot]{save_nfdplot()}}
}
