% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_year_type_get.R
\name{sac_year_type_get}
\alias{sac_year_type_get}
\title{Get Sacramento Year Type Index Data}
\usage{
sac_year_type_get(internal = FALSE)
}
\arguments{
\item{internal}{Boolean. If \code{TRUE}, will use the internal data saved with
the package instead of downloading the data.}
}
\value{
\code{xts} object with 1 column.
}
\description{
\code{sac_year_type_get()} gets the historical record of Sacramento Year Type
indexes, which are required input to CRSS. The function will download them
from http://cdec.water.ca.gov/cgi-progs/iodir/WSIHIST and then parse the
required information. If an internet connection is not available, then the
internal data can be used, but it may not be up-to-date.
}
\details{
The "crssio.sac_yt_url" option determines where the function downloads the
data from.
}
