% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_boxplot_custom.R
\docType{data}
\name{stat_boxplot_custom}
\alias{stat_boxplot_custom}
\alias{StatBoxplotCustom}
\title{A box and whiskers plot that uses percentiles instead of IQR}
\usage{
stat_boxplot_custom(
  mapping = NULL,
  data = NULL,
  geom = "boxplot",
  position = "dodge",
  ...,
  qs = c(0.05, 0.25, 0.5, 0.75, 0.95),
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{Use to override the default connection between
\code{geom_boxplot} and \code{stat_boxplot}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{qs}{The percentiles that are used to create the lower whisker, lower
hinge, middle bar, upper hinge, and upper whisker, respectively. The lower
and upper whiskers default to the 5th and 95th percentiles. The hinges
default to the 25th and 75th percentiles, and the middle bar defaults to
the median. These  values should be in increasing order, i.e., the whisker
should be a smaller percentile than the lower hinge, and can only span the
values from 0 to 1 (inclusive).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA})
automatically determines the orientation from the aesthetic mapping. In the
rare event that this fails it can be given explicitly by setting \code{orientation}
to either \code{"x"} or \code{"y"}. See the \emph{Orientation} section for more detail.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
\code{stat_boxplot_custom()} modifies \code{\link[ggplot2:geom_boxplot]{ggplot2::stat_boxplot()}} so that it
computes the extents of the whiskers based on specified percentiles, rather
than a multiple of the IQR.
}
\details{
The upper and lower whiskers extend to the first, and last entries of the
\code{qs} parameter, respectively. Data beyond the whiskers are "outliers".
}
\examples{

library(ggplot2)
p <- ggplot(mpg, aes(class, hwy))

# show whiskers at 5th and 95th percentiles
p + stat_boxplot_custom(qs = c(.05, .25, .5, .75, .95))

# show whiskers at 10th and 90th percentiles
p + stat_boxplot_custom(qs = c(.1, .25, .5, .75, .9))

}
\keyword{datasets}
